requirements.txt
flask
requests
python-dotenv
pinecone
sentence-transformers
flask-cors